// packages/engine/src/assets/enums/AssetType.d.ts
export declare const AssetType: {
    glTF: number;
    FBX: number;
    OBJ: number;
    VRM: number;
    PNG: number;
    JPEG: number;
    MP4: number;
    TS: number;
    MKV: number;
    MP3: number;
    OGG: number;
    AAC: number;
    CSV: number;
    PlainText: number;
    DOC: number;
    XLS: number;
    Script: number;
};